#225 model saving now works

+++ src/core/db3/server/db3mutationCore.ts
@@ -8,7 +8,7 @@ import { nanoid } from 'nanoid';
 import { ComputeChangePlan } from "shared/associationUtils";
 import { Permission } from "shared/permissions";
 import { DateTimeRange } from "shared/time";
-import { ChangeAction, ChangeContext, CreateChangeContext, ObjectDiff, RegisterChange, getIntersectingFields, sanitize } from "shared/utils";
+import { ChangeAction, ChangeContext, CoalesceBool, CreateChangeContext, ObjectDiff, RegisterChange, getIntersectingFields, sanitize } from "shared/utils";
 import sharp from "sharp";
 import * as db3 from "../db3";
 import { CMDBTableFilterModel, FileCustomData, ForkImageParams, ImageFileFormat, ImageMetadata, TAnyModel, TinsertOrUpdateEventSongListArgs, TinsertOrUpdateEventSongListDivider, TinsertOrUpdateEventSongListSong, TupdateEventCustomFieldValue, TupdateEventCustomFieldValuesArgs, WorkflowObjectType, getFileCustomData } from "../shared/apiTypes";
@@ -1253,7 +1253,15 @@ export type SyncEntitiesResult = {
 //
 // one limitation of this function is that it cannot deal with self-referencing entities. for example,
 // if a node entity contains a "nextNodeId", and both have provisional ids, then we would need extra logic to order in a DAG, ensure creating in a safe order, and updating from provisional to real IDs each creation.
-export async function SyncNonSelfReferencingEntities<T extends { id: number }>(
+export async function SyncNonSelfReferencingEntities<T extends { id: number }>({
+    entityName,
+    existingEntities,
+    desiredEntities,
+    allowedKeysForCreate,
+    dbOperations,
+    ignoreDiffFieldsForUpdates,
+    options,
+}: {
     entityName: string,
     existingEntities: T[],
     desiredEntities: T[],
@@ -1263,8 +1271,14 @@ export async function SyncNonSelfReferencingEntities<T extends { id: number }>(
         update: (id: number, data: Partial<T>) => Promise<any>,
         create: (data: Omit<T, "id">) => Promise<T>,
     },
+    options?: {
+        allowDeletions?: boolean | undefined, // default true
+    },
     ignoreDiffFieldsForUpdates?: (keyof Omit<T, "id">)[], // only for updates, fields of T which should not ever be updated or contribute to equality check. "id" is automatically included
-): Promise<SyncEntitiesResult> {
+}): Promise<SyncEntitiesResult> {
+
+    const allowDeletions = CoalesceBool(options?.allowDeletions, true);
+
     // Compute change plan
     const changePlan = ComputeChangePlan(existingEntities, desiredEntities, (a, b) => a.id === b.id);
     const result: SyncEntitiesResult = {
@@ -1273,15 +1287,17 @@ export async function SyncNonSelfReferencingEntities<T extends { id: number }>(
     };
 
     // Delete
-    const idsToDelete = changePlan.delete.map(x => x.id);
-    if (idsToDelete.length > 0) {
-        await dbOperations.deleteMany(idsToDelete);
-        result.changes.push(...changePlan.delete.map(x => ({
-            action: ChangeAction.delete,
-            pkid: x.id,
-            objectType: entityName as WorkflowObjectType,
-            oldValues: x,
-        })));
+    if (allowDeletions) {
+        const idsToDelete = changePlan.delete.map(x => x.id);
+        if (idsToDelete.length > 0) {
+            await dbOperations.deleteMany(idsToDelete);
+            result.changes.push(...changePlan.delete.map(x => ({
+                action: ChangeAction.delete,
+                pkid: x.id,
+                objectType: entityName as WorkflowObjectType,
+                oldValues: x,
+            })));
+        }
     }
 
     // Update
